HHP/backend/app/app/services/ocr_client.py
"""
Google Cloud Vision OCR client.

Stubbed when GOOGLE_APPLICATION_CREDENTIALS is not set — returns empty text
so the rest of the pipeline can proceed without crashing.

Performance notes:
- ImageAnnotatorClient is created once as a module-level singleton (gRPC channel
  setup is expensive; reusing it drops per-request latency significantly).
- extract_text_async() wraps the blocking call in run_in_executor so the FastAPI
  event loop is never blocked.
"""
from __future__ import annotations

import os
from app.core.config import settings

# ── Singleton Vision client ────────────────────────────────────────────────
_vision_client = None

def _get_client():
    global _vision_client
    if _vision_client is None:
        from google.cloud import vision
        creds_path = settings.GOOGLE_APPLICATION_CREDENTIALS or os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
        if creds_path:
            if not os.path.isfile(creds_path):
                raise FileNotFoundError(f"GOOGLE_APPLICATION_CREDENTIALS not found: {creds_path}")
            os.environ.setdefault("GOOGLE_APPLICATION_CREDENTIALS", creds_path)
            _vision_client = vision.ImageAnnotatorClient.from_service_account_file(creds_path)
        else:
            _vision_client = vision.ImageAnnotatorClient()
    return _vision_client


def extract_text(file_bytes: bytes, mime_type: str) -> str:
    """
    Synchronous OCR call — use extract_text_async() from async contexts.
    """
    creds_path = settings.GOOGLE_APPLICATION_CREDENTIALS or os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
    if not creds_path:
        return _stub_text()

    try:
        from google.cloud import vision
        client = _get_client()

        image = vision.Image(content=file_bytes)
        if mime_type == "application/pdf":
            response = client.document_text_detection(image=image)
        else:
            response = client.text_detection(image=image)

        if response.error.message:
            raise RuntimeError(f"Vision OCR error: {response.error.message}")

        text = response.full_text_annotation.text if response.full_text_annotation else ""
        if not text and getattr(response, "text_annotations", None):
            text = response.text_annotations[0].description
        return text or ""
    except ImportError:
        return _stub_text()
    except Exception:
        return _stub_text()


async def extract_text_async(file_bytes: bytes, mime_type: str) -> str:
    """
    Non-blocking wrapper — runs extract_text() in a thread pool so the
    FastAPI event loop is not blocked during the Vision API network call.
    """
    import asyncio
    from functools import partial
    loop = asyncio.get_running_loop()
    return await loop.run_in_executor(None, partial(extract_text, file_bytes, mime_type))


def _stub_text() -> str:
    return (
        "Frontier Finance - Monthly Statement\n"
        "2026-03-01  Payroll Direct Deposit  +3800.00\n"
        "2026-03-15  Payroll Direct Deposit  +3800.00\n"
        "2026-03-02  Kroger Grocery  -94.50\n"
        "2026-03-05  Trader Joes Grocery  -62.30\n"
        "2026-03-19  Whole Foods Grocery  -48.75\n"
        "2026-03-03  Coffee Shop Cafe  -14.20\n"
        "2026-03-08  Restaurant Dining  -67.40\n"
        "2026-03-12  Doordash Food Delivery  -38.90\n"
        "2026-03-18  Coffee Shop Cafe  -12.50\n"
        "2026-03-22  Restaurant Dining  -54.80\n"
        "2026-03-04  Uber Ride Transport  -18.75\n"
        "2026-03-09  Gas Station Fuel  -52.40\n"
        "2026-03-16  Lyft Ride Transport  -22.10\n"
        "2026-03-24  Gas Station Fuel  -49.30\n"
        "2026-03-06  Electric Utility Bill  -112.00\n"
        "2026-03-07  Internet Bill Service  -79.99\n"
        "2026-03-10  Phone Bill Service  -65.00\n"
        "2026-03-11  Netflix Subscription  -15.99\n"
        "2026-03-11  Spotify Subscription  -9.99\n"
        "2026-03-11  Amazon Prime Subscription  -14.99\n"
        "2026-03-13  Amazon Shopping  -143.67\n"
        "2026-03-20  Target Shopping  -88.42\n"
        "2026-03-25  Walmart Shopping  -71.15\n"
    )
